requirements.txt
# Connexion + Swagger tools
connexion[swagger-ui]==2.7.0
Werkzeug==2.0.1
PyYAML==5.4.1
prance==0.21.8.0

# General utils
requests==2.25.1
six==1.16.0
Jinja2==3.0.1

# JWT / auth
python-jose[cryptography]==3.3.0
bcrypt==3.2.0

# Kubernetes
kubernetes==19.15.0a1

# MongoDB
pymongo==4.0.1

# Deprecated: etcd (use mongo instead)
python-etcd==0.4.5

# RabbitMQ
rabbitpy==2.0.1

# Manual validation?
# jsonschema==3.2.0
# openapi2jsonschema==0.9.1